+++ examples/server/a_bit_of_everything.go
@@ -6,6 +6,7 @@ import (
 	"sync"
 
 	"github.com/golang/glog"
+	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes/duration"
 	"github.com/golang/protobuf/ptypes/empty"
 	examples "github.com/grpc-ecosystem/grpc-gateway/examples/examplepb"
@@ -13,6 +14,7 @@ import (
 	sub2 "github.com/grpc-ecosystem/grpc-gateway/examples/sub2"
 	"github.com/rogpeppe/fastuuid"
 	"golang.org/x/net/context"
+	"google.golang.org/genproto/googleapis/rpc/errdetails"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
@@ -245,4 +247,22 @@ func (s *_ABitOfEverythingServer) Timeout(ctx context.Context, msg *empty.Empty)
 	case <-ctx.Done():
 		return nil, ctx.Err()
 	}
+}
+
+func (s *_ABitOfEverythingServer) ErrorWithDetails(ctx context.Context, msg *empty.Empty) (*empty.Empty, error) {
+	stat, err := status.New(codes.Unknown, "with details").
+		WithDetails(proto.Message(
+			&errdetails.DebugInfo{
+				StackEntries: []string{"foo:1"},
+				Detail:       "error debug details",
+			},
+		))
+	if err != nil {
+		return nil, status.Errorf(codes.Internal, "unexpected error adding details: %s", err)
+	}
+	return nil, stat.Err()
+}
+
+func (s *_ABitOfEverythingServer) GetMessageWithBody(ctx context.Context, msg *examples.MessageWithBody) (*empty.Empty, error) {
+	return &empty.Empty{}, nil
 }