+++ protoc-gen-grpc-gateway/descriptor/types.go
@@ -9,6 +9,12 @@ import (
 	"github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/httprule"
 )
 
+// IsWellKnownType returns true if the provided fully qualified type name is considered 'well-known'.
+func IsWellKnownType(typeName string) bool {
+	_, ok := wellKnownTypeConv[typeName]
+	return ok
+}
+
 // GoPackage represents a golang package
 type GoPackage struct {
 	// Path is the package path to the package.
@@ -195,6 +201,9 @@ func (p Parameter) ConvertFuncExpr() (string, error) {
 	typ := p.Target.GetType()
 	conv, ok := tbl[typ]
 	if !ok {
+		conv, ok = wellKnownTypeConv[p.Target.GetTypeName()]
+	}
+	if !ok {
 		return "", fmt.Errorf("unsupported field type %s of parameter %s in %s.%s", typ, p.FieldPath, p.Method.Service.GetName(), p.Method.GetName())
 	}
 	return conv, nil
@@ -286,8 +295,7 @@ var (
 		descriptor.FieldDescriptorProto_TYPE_STRING:  "runtime.String",
 		// FieldDescriptorProto_TYPE_GROUP
 		// FieldDescriptorProto_TYPE_MESSAGE
-		// FieldDescriptorProto_TYPE_BYTES
-		// TODO(yugui) Handle bytes
+		descriptor.FieldDescriptorProto_TYPE_BYTES:  "runtime.Bytes",
 		descriptor.FieldDescriptorProto_TYPE_UINT32: "runtime.Uint32",
 		// FieldDescriptorProto_TYPE_ENUM
 		// TODO(yugui) Handle Enum
@@ -318,5 +326,10 @@ var (
 		descriptor.FieldDescriptorProto_TYPE_SFIXED64: "runtime.Int64P",
 		descriptor.FieldDescriptorProto_TYPE_SINT32:   "runtime.Int32P",
 		descriptor.FieldDescriptorProto_TYPE_SINT64:   "runtime.Int64P",
+	}
+
+	wellKnownTypeConv = map[string]string{
+		".google.protobuf.Timestamp": "runtime.Timestamp",
+		".google.protobuf.Duration":  "runtime.Duration",
 	}
 )