+++ protoc-gen-grpc-gateway/main.go
@@ -24,7 +24,8 @@ import (
 
 var (
 	importPrefix      = flag.String("import_prefix", "", "prefix to be added to go package paths for imported proto files")
-	useRequestContext = flag.Bool("request_context", false, "determine whether to use http.Request's context or not")
+	importPath        = flag.String("import_path", "", "used as the package if no input files declare go_package. If it contains slashes, everything up to the rightmost slash is ignored.")
+	useRequestContext = flag.Bool("request_context", true, "determine whether to use http.Request's context or not")
 	allowDeleteBody   = flag.Bool("allow_delete_body", false, "unless set, HTTP DELETE methods may not have a body")
 )
 
@@ -78,6 +79,7 @@ func main() {
 	g := gengateway.New(reg, *useRequestContext)
 
 	reg.SetPrefix(*importPrefix)
+	reg.SetImportPath(*importPath)
 	reg.SetAllowDeleteBody(*allowDeleteBody)
 	if err := reg.Load(req); err != nil {
 		emitError(err)