+++ runtime/errors.go
@@ -63,11 +63,12 @@ var (
 )
 
 type errorBody struct {
-	Error string `protobuf:"bytes,1,name=error" json:"error"`
-	Code  int32  `protobuf:"varint,2,name=code" json:"code"`
+	Error   string          `protobuf:"bytes,1,name=error" json:"error"`
+	Code    int32           `protobuf:"varint,2,name=code" json:"code"`
+	Details []proto.Message `protobuf:"bytes,3,name=details" json:"details"`
 }
 
-//Make this also conform to proto.Message for builtin JSONPb Marshaler
+// Make this also conform to proto.Message for builtin JSONPb Marshaler
 func (e *errorBody) Reset()         { *e = errorBody{} }
 func (e *errorBody) String() string { return proto.CompactTextString(e) }
 func (*errorBody) ProtoMessage()    {}
@@ -91,8 +92,11 @@ func DefaultHTTPError(ctx context.Context, mux *ServeMux, marshaler Marshaler, w
 		Code:  int32(s.Code()),
 	}
 
-	w.Header().Del("Trailer")
-	w.Header().Set("Content-Type", marshaler.ContentType(body))
+	for _, detail := range s.Details() {
+		if det, ok := detail.(proto.Message); ok {
+			body.Details = append(body.Details, det)
+		}
+	}
 
 	buf, merr := marshaler.Marshal(body)
 	if merr != nil {